Do not dereference symlinks when installing
It'd be nice if install(1) could do this. I could probably just use
cp(1) for everything. Closes #62

+++ install.sh
@@ -125,6 +125,10 @@ for f in *; do
     fi
 
     # install
-    install $verbose $force -t "$DESTDIR" -- "$f"
+    if ! [ -L "$f" ]; then
+        install $verbose $force -t "$DESTDIR" -- "$f"
+    else
+        cp $verbose $force -d --preserve=all "$f" "$DESTDIR"
+    fi
 done
